bin/server_tail_requirements.txt
python-sshtail==0.0.2
termcolor==1.1.0
